data type and mapping orm to domain model

+++ domain/model/item.go
@@ -2,18 +2,18 @@ package model
 
 // Items can be goods, products, gifts, services...
 type Item struct {
-	Code         string `json:"item_code"`
-	Name         string `json:"item_name"`
-	Description  string `json:"item_description"`
-	BrandCode    string `json:"brand_code"`
-	Image        string `json:"item_image"`
-	CategoryCode string `json:"category_code"`
-	ListedPrice  int32  `json:"listed_price"`
-	SalesPrice   int32  `json:"sales_price"`
-	FullPrice    int32  `json:"full_price"`
-	Type         string `json:"type"`
-	IsActived    int8   `json:"is_actived"`
-	IsDeleted    int8   `json:"is_deleted"`
+	Code         string  `json:"item_code"`
+	Name         string  `json:"item_name"`
+	Description  string  `json:"item_description"`
+	BrandCode    string  `json:"brand_code"`
+	Image        string  `json:"item_image"`
+	CategoryCode string  `json:"category_code"`
+	ListedPrice  float32 `json:"listed_price"`
+	SalesPrice   float32 `json:"sales_price"`
+	FullPrice    float32 `json:"full_price"`
+	Type         string  `json:"type"`
+	IsActived    int8    `json:"is_actived"`
+	IsDeleted    int8    `json:"is_deleted"`
 
 	// The key-value collection of variant attributes. Ex: color, size...
 	// ItemAttribute *map[string]string `json:"item_attribute"`
